Move schedule indicator to spent column
